+++ src/main/java/io/github/torand/fastersql/setoperation/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Provides classes representing set operations.
+ */
+package io.github.torand.fastersql.setoperation;