+++ src/test/java/io/github/torand/fastersql/statement/sqlserver/SqlServerSelectStatementTest.java
@@ -17,6 +17,7 @@ package io.github.torand.fastersql.statement.sqlserver;
 
 import io.github.torand.fastersql.datamodel.CustomerTable;
 import io.github.torand.fastersql.statement.PreparableStatement;
+import io.github.torand.fastersql.statement.SelectSetOpStatement;
 import io.github.torand.fastersql.statement.SelectStatement;
 import org.junit.jupiter.api.Test;
 
@@ -49,7 +50,6 @@ import static io.github.torand.fastersql.function.singlerow.SingleRowFunctions.t
 import static io.github.torand.fastersql.function.singlerow.SingleRowFunctions.upper;
 import static io.github.torand.fastersql.function.system.SystemFunctions.currentDate;
 import static io.github.torand.fastersql.function.system.SystemFunctions.currentTimestamp;
-import static io.github.torand.fastersql.predicate.compound.CompoundPredicates.not;
 import static io.github.torand.fastersql.projection.Projections.colPos;
 import static io.github.torand.fastersql.projection.Projections.subquery;
 import static io.github.torand.fastersql.relation.Relations.table;
@@ -77,7 +77,7 @@ public class SqlServerSelectStatementTest extends SqlServerTest {
                 .join(PURCHASE.ID.on(PURCHASE_ITEM.PURCHASE_ID))
                 .join(PURCHASE_ITEM.PRODUCT_ID.on(PRODUCT.ID))
                 .where(CUSTOMER.MOBILE_NO_VERIFIED.eq(true)
-                    .and(not(CUSTOMER.EMAIL_ADDRESS.isNull()))
+                    .and(CUSTOMER.EMAIL_ADDRESS.isNotNull())
                     .and(CUSTOMER.COUNTRY_CODE.in("NOR", "DEN"))
                     .and(PRODUCT.CATEGORY.eq("FURNITURE").or(PRODUCT.CATEGORY.eq("LAMP")))
                     .and(PURCHASE.CREATED_TIME.gt(since)))
@@ -573,6 +573,60 @@ public class SqlServerSelectStatementTest extends SqlServerTest {
                 "DIVIDE_", isBigDecimalCloseTo(1358.375, 0.0001),
                 "MOD_", isBigDecimalCloseTo(3.50, 0.001),
                 "NEG_", isInteger(-13))
+            .verify(stmt);
+    }
+
+    @Test
+    void shouldHandleSetOperations() {
+        // SQL Server does not support ORDER BY in set operation operands and no parentheses around operands
+        SelectSetOpStatement stmt =
+            select(PRODUCT.NAME.as("PR_NAME"))
+                .from(PRODUCT)
+                .where(PRODUCT.NAME.like("Apple").or(PRODUCT.NAME.like("Ekornes")))
+                .intersect(
+                    select(PRODUCT.NAME)
+                        .from(PRODUCT)
+                        .where(PRODUCT.NAME.like("Ekornes"))
+                )
+                .unionAll(
+                    select(PRODUCT.NAME)
+                        .from(PRODUCT)
+                        .where(PRODUCT.NAME.like("Apple").or(PRODUCT.NAME.like("Electrolux")))
+                )
+                .except(
+                    select(PRODUCT.NAME)
+                        .from(PRODUCT)
+                        .where(PRODUCT.NAME.like("Electrolux"))
+                )
+                .orderBy(alias("PR_NAME").asc());
+
+        statementTester()
+            .assertSql("""
+                select PR.NAME PR_NAME \
+                from PRODUCT PR \
+                where PR.NAME like ? or PR.NAME like ? \
+                intersect \
+                select PR.NAME PR_NAME \
+                from PRODUCT PR \
+                where PR.NAME like ? \
+                union all \
+                select PR.NAME PR_NAME \
+                from PRODUCT PR \
+                where PR.NAME like ? or PR.NAME like ? \
+                except \
+                select PR.NAME PR_NAME \
+                from PRODUCT PR \
+                where PR.NAME like ? \
+                order by PR_NAME asc"""
+            )
+            .assertParams("%Apple%", "%Ekornes%", "%Ekornes%", "%Apple%", "%Electrolux%", "%Electrolux%")
+            .assertRowCount(2)
+            .assertRow(1,
+                "PR_NAME", is("Apple iPad Pro tablet")
+            )
+            .assertRow(2,
+                "PR_NAME", is("Ekornes Stressless resting chair")
+            )
             .verify(stmt);
     }
 }